Rename kubeconfig test to workload cluster (#659)
